+++ src/Plumix/Rendering/DecoratedSliver.cs
@@ -73,7 +73,7 @@ public sealed class RenderDecoratedSliver : RenderProxySliver
 
     public override void Paint(PaintingContext context, Point offset)
     {
-        if (Child == null || Geometry.PaintExtent <= 0.0)
+        if (Child == null || !Geometry.Visible)
         {
             return;
         }